+++ scripts/release/generate_external_installer_assets.sh
@@ -24,7 +24,10 @@ function exitWithUsage {
   echo "    │  ├─ recalbox-rpi3.img.xz"
   echo "    │  └─ boot.tar.xz"
   echo "    └─ rpi4/"
-  echo "       ├─ recalbox-rpi4.img.xz"
+  echo "    │  ├─ recalbox-rpi4.img.xz"
+  echo "    │  └─ boot.tar.xz"
+  echo "    ├─ rpizero2legacy/"
+  echo "       ├─ recalbox-rpizero2legacy.img.xz"
   echo "       └─ boot.tar.xz"
   echo
   exit 64
@@ -86,7 +89,7 @@ function generateRaspberryPiImagerAssets {
   metadata[version]=${CI_COMMIT_REF_NAME}
   metadata[releaseDate]=$(date +%Y-%m-%d)
 
-  for arch in rpi1 rpi2 rpi3 rpi4; do
+  for arch in rpi1 rpi2 rpi3 rpi4 rpizero2legacy; do
     local imageFile="${params[imagesDir]}/${arch}/recalbox-${arch}.img.xz"
     # Fetch info regarding image downloads (XZ-compressed Recalbox image)
     metadata["${arch}ImageDownloadSize"]=$(stat -c '%s' "${imageFile}")
@@ -104,18 +107,22 @@ function generateRaspberryPiImagerAssets {
   | sed -e "s|{{version}}|${metadata[version]}|" \
         -e "s|{{releaseDate}}|${metadata[releaseDate]}|" \
         -e "s|{{rpi1ExtractSize}}|${metadata[rpi1ExtractSize]}|" \
+        -e "s|{{rpizero2legacyExtractSize}}|${metadata[rpizero2legacyExtractSize]}|" \
         -e "s|{{rpi2ExtractSize}}|${metadata[rpi2ExtractSize]}|" \
         -e "s|{{rpi3ExtractSize}}|${metadata[rpi3ExtractSize]}|" \
         -e "s|{{rpi4ExtractSize}}|${metadata[rpi4ExtractSize]}|" \
         -e "s|{{rpi1ExtractSha256}}|${metadata[rpi1ExtractSha256]}|" \
+        -e "s|{{rpizero2legacyExtractSha256}}|${metadata[rpizero2legacyExtractSha256]}|" \
         -e "s|{{rpi2ExtractSha256}}|${metadata[rpi2ExtractSha256]}|" \
         -e "s|{{rpi3ExtractSha256}}|${metadata[rpi3ExtractSha256]}|" \
         -e "s|{{rpi4ExtractSha256}}|${metadata[rpi4ExtractSha256]}|" \
         -e "s|{{rpi1ImageDownloadSize}}|${metadata[rpi1ImageDownloadSize]}|" \
+        -e "s|{{rpizero2legacyImageDownloadSize}}|${metadata[rpizero2legacyImageDownloadSize]}|" \
         -e "s|{{rpi2ImageDownloadSize}}|${metadata[rpi2ImageDownloadSize]}|" \
         -e "s|{{rpi3ImageDownloadSize}}|${metadata[rpi3ImageDownloadSize]}|" \
         -e "s|{{rpi4ImageDownloadSize}}|${metadata[rpi4ImageDownloadSize]}|" \
         -e "s|{{rpi1ImageDownloadSha256}}|${metadata[rpi1ImageDownloadSha256]}|" \
+        -e "s|{{rpizero2legacyImageDownloadSha256}}|${metadata[rpizero2legacyImageDownloadSha256]}|" \
         -e "s|{{rpi2ImageDownloadSha256}}|${metadata[rpi2ImageDownloadSha256]}|" \
         -e "s|{{rpi3ImageDownloadSha256}}|${metadata[rpi3ImageDownloadSha256]}|" \
         -e "s|{{rpi4ImageDownloadSha256}}|${metadata[rpi4ImageDownloadSha256]}|" \